Make a whole bunch of methods private
I just found out why so many methods in Odoo models starts with underscore.
I thought it was just a matter of style. Nope. Methods starting with
underscore are not exposed via the public XML-RPC API.
I went back and changed names of a bunch of our methods that doesn't
need to be accessible via the RPC API.